Add team controller

+++ wabit/src/main/java/com/springboot/wabit/security/JWTAuthenticationFilter.java
@@ -62,7 +62,7 @@ public class JWTAuthenticationFilter extends UsernamePasswordAuthenticationFilte
         String token = Jwts
                 .builder()
                 .setSubject(username)
-                .setExpiration(new Date(System.currentTimeMillis() + EXPIRATION_TIME))
+                .setExpiration(new Date((long) (System.currentTimeMillis() + EXPIRATION_TIME)))
                 .signWith(SignatureAlgorithm.HS512, SECRET)
                 .compact();
         String bearerToken = TOKEN_PREFIX + token;
